+++ tests/testthat/test-bootspat_naive.R
@@ -11,6 +11,8 @@ test_that("function bootspat_naive works", {
 
   # loading data
   bin1 <- load_ext_data() #terra::rast(system.file("extdata", "spp_sites.tif", package="SESraster"))
+  bin1[1] <- NA
+  bin1[2] <- 0
 
   # applying the function
   rand.site <- bootspat_naive(bin1, "site")
@@ -25,9 +27,13 @@ test_that("function bootspat_naive works", {
   expect_true(inherits(rand.sp, "SpatRaster"), "TRUE")
   expect_true(inherits(rand.both, "SpatRaster"), "TRUE")
 
-  expect_equal(unlist(rand.site[2]), setNames(c(0,0,1,0,1), names(bin1)))
+  expect_equal(unlist(rand.site[2]), setNames(c(0,0,0,0,0), names(bin1)))
   expect_equal(unlist(rand.sp[2]), setNames(c(0,1,1,0,0), names(bin1)))
   expect_equal(unlist(rand.both[2]), setNames(c(1,0,0,0,1), names(bin1)))
 
-  expect_equal(sum(rand.site)[1:8], sum(bin1)[1:8])
+  expect_equal(unlist(rand.site[1]), setNames(as.double(rep(NA, terra::nlyr(bin1))), names(bin1)))
+  expect_equal(unlist(rand.sp[1]), setNames(as.double(rep(NA, terra::nlyr(bin1))), names(bin1)))
+  expect_equal(unlist(rand.both[1]), setNames(as.double(rep(NA, terra::nlyr(bin1))), names(bin1)))
+
+  expect_equal(sum(rand.site)[1:8], sum(bin1)[1:8]) # test richness
 })