+++ NEWS.md
@@ -1,7 +1,9 @@
 # SESraster 0.6.2
 
 ## enhancements
-- improved accuracy of algorithm description in DESCRIPTION, README, and vignette
+- added new vignette
+- added references and improved accuracy of algorithm description in DESCRIPTION, README, and vignettes
+- improved accuracy of null model descriptions in vignettes
 - added link to functions in documentation
 
 # SESraster 0.6.1
@@ -9,4 +11,5 @@
 Released 2023-06-23
 
 * Initial CRAN submission.
+