+++ NEWS.md
@@ -1,5 +1,8 @@
 # SESraster 0.6.2
 
+## bux fixes
+- fixed error when all species were absent from a cell (i.e. all values are zero)
+
 ## enhancements
 - added new vignette
 - added references and improved accuracy of algorithm description in DESCRIPTION, README, and vignettes